backend/pyproject.toml
[project]
name = "hackathon-todo-cli"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.1",
    "mcp>=1.26.0",
    "openai-agents>=0.8.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "requests>=2.32.5",
    "sqlmodel>=0.0.32",
    "uvicorn>=0.40.0",
]
